@@ -28,8 +28,8 @@
     "leva": "^0.9.29",
     "lodash.isfinite": "^3.3.2",
     "process": "^0.11.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.6",
     "redux": "^4.2.0",
     "redux-logger": "^3.0.6",
